requirements.txt
jieba==0.42.1
Scrapy==2.8.0
wordcloud==1.8.2.2
python-dotenv==1.0.0
